requirements.txt
torch>=2.0
torch_geometric>=2.5
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.13
rasterio>=1.3
streamlit>=1.30
plotly>=5.18
pillow>=10.0
